Extract code from AsHtml into Render
Render now includes a new type, RenderedModule, and exports a function
renderModule, which takes care of:

* ensuring only exported values appear
* For those declarations which should appear in the docs, extracting a
  title for each

This means that the html-specific code no longer needs to worry about
these things.

This code also adds a new requirement: that all declarations which will
appear in the docs have to have a title. For now, I've re-added titles
for instances. This will probably be changed later (see #15).