Better rendering of instances #15
* Create a new RenderedChildDeclaration type for declarations which
  should appear under other declarations. At the moment, this is only
  type class members, type class instances, and data constructors.
* Move RenderedChildDeclaration values under their respective parents
  in a new second pass in PscPages.Render.

These changes are also related to purescript/purescript#747, and
they do even fix some occasions where that bug occurs, although not all.
For example, if a module defines and exports a type class, and
additionally defines but does not export a type which is an instance of
the type class, its instance will appear in the rendered module, but
should not.